# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ database/migrations/2021_04_27_193316_create_kitchen_stocks_table.php
@@ -17,7 +17,7 @@ class CreateKitchenStocksTable extends Migration
             $table->id();
             $table->bigInteger('material_id')->unsigned();
                 $table->foreign('material_id')->references('id')->on('materials');
-            $table->integer('quantity')->unsigned();
+            $table->decimal('quantity')->unsigned();
             $table->softDeletes();
             $table->timestamps();
         });
